main: refactor StringDecoder to be used as a library
reworked a bunch of the main script to make it easy for FLOSS to be used as a library for client programs. the work is not done yet. the major task was extract the parsing of cli arguments from the StringDecoder class into the main script, and provide these values as parameters to the relevant functions. this ultimately removed the Config class.

I moved output functions off the main classes, since its presumptuous for us to know how a client program would want to deal with the extracted strings. instead, these routines are now functions. this ends up making sense because the routines simply format the data passed in, and do not need extra context provided by a class.